+++ src/getEsbuildConfig.ts
@@ -1,20 +1,22 @@
 import validateEntryPoints from 'validateEntryPoints'
 import parseOptions from 'parseOptions'
 import type { BuildOptions } from 'esbuild'
-import type { CLIFlags } from 'types'
+import getCLIOptions from 'getOptions'
 
-type GetEsbuildConfig = (options: CLIFlags) => Promise<BuildOptions>
+type GetEsbuildConfig = () => Promise<BuildOptions>
 
-const getEsbuildConfig: GetEsbuildConfig = async options => {
+const getEsbuildConfig: GetEsbuildConfig = async () => {
+  const options = getCLIOptions()
   const entryPoints = await validateEntryPoints(options)
-  const { external, outExtension, outdir, splitting } = await parseOptions(options)
+  const { external, incremental, outExtension, outdir, splitting } = await parseOptions()
   return {
     bundle: options.bundle,
     entryPoints,
-    outExtension,
     format: options.format,
+    incremental,
     minify: options.minify,
     outdir,
+    outExtension,
     platform: options.platform,
     ...(options.bundle && {
       external,