+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
-#### [Unreleased](https://github.com/henriquehbr/eswatch/compare/v0.10.0...HEAD)
+#### [v0.10.1](https://github.com/henriquehbr/eswatch/compare/v0.10.0...v0.10.1)
 
+- Separate core and helpers modules [`#3`](https://github.com/henriquehbr/eswatch/pull/3)
 - refactor: modularize watch/build process [`96943a5`](https://github.com/henriquehbr/eswatch/commit/96943a53e7788bf4bcdec5ddb67415b103142aa6)
 - refactor: modularize CLI options parsing [`50be3b7`](https://github.com/henriquehbr/eswatch/commit/50be3b7223dfaac175ad7df164c81e4ea32f9de4)
 - refactor: return build steps through Promise.all() [`f4f9b2b`](https://github.com/henriquehbr/eswatch/commit/f4f9b2b818754f63b38a777e971f1946d804104e)
@@ -12,7 +13,7 @@ All notable changes to this project will be documented in this file. Dates are d
 - refactor(parseOptions): add "incremental" option [`3b60fc3`](https://github.com/henriquehbr/eswatch/commit/3b60fc3e94464c452f2e23a2d83cf896e6e7ce8a)
 - chore: bump dependencies [`f183e10`](https://github.com/henriquehbr/eswatch/commit/f183e10b058040f5d777209bc9985bfadbb44386)
 - refactor: redirect imports to helpers to barrel file [`aaebc7c`](https://github.com/henriquehbr/eswatch/commit/aaebc7c87a874b525c85423c7a692ea8284abd81)
-- refactor(validateEntryPoints): move to helpers [`47ba633`](https://github.com/henriquehbr/eswatch/commit/47ba63375df36cf81bd98871d6769fa231cee7cb)
+- refactor(validateEntryPoints): move to helpers [`48172e8`](https://github.com/henriquehbr/eswatch/commit/48172e8037dcb7ee337f6484b62439f693b5dff1)
 
 #### [v0.10.0](https://github.com/henriquehbr/eswatch/compare/v0.9.5...v0.10.0)